Chain Page Refactoring
