+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "pooltogether-stats",
-	"version": "2.5.2",
+	"version": "2.6.0",
 	"type": "module",
 	"private": true,
 	"scripts": {
